finished memory documentation

+++ include/rud/memory.hpp
@@ -36,7 +36,7 @@ namespace rud {
 
     void mem_copy(void* dest, const void* src, u64 size);
     void mem_move(void* dest, const void* src, u64 size);
-    void mem_set(void* dest, const u32 value, u64 size);
+    void mem_set(void* dest, const u8 value, u64 size);
     bool mem_equals(const void* ptr1, const void* ptr2, u64 size);
 }
